Explain hybrid encryption
When defining your delay encryption API in section 2, it may be
found helpful to have it followed by a sentence explaining how to
use the four algorithms for the purpose of, e.g., the Vickrey
auction from the introduction. Now we only have a key k and an
encryption c of it, so a line saying that k can then be used to
encrypt a message m using some symmetric scheme (as on p7) could
help the reader connecting the dots.