Remove a misleading statement
The sentences "In particular, isogenies preserve the group
order of elliptic curves, and thus they preserve singularity." is
imprecise, and not only because the kernel may not be defined over
the base field. E.g., the isomorphism from a curve to its
quadratic twist has trivial kernel, yet it does not need to
preserve the group order. Maybe the simplest thing is just to
declare that isogenies preserve supersingularity, as a known fact,
and not try to explain it using the group order.