Nitpicking
p8: "The full supersingular graph of GF(p), ..." --> do you mean
GF(p^2), or perhaps the algebraic closure?

p11 in the statement of Theorem 11: would it make sense to replace
"efficient" with poly(lambda, T)? That's not really the same as
"efficient".

p12 in the proof of Theorem 11, maybe write somewhere that one can
treat Q as synonymous to id, which is what you do