Added contributions, other touches to intro
A minor point is that they authors state that the only
construction of fully homomorphic time-lock puzzles require
obfuscation, whereas we know how to construct them from lattices plus
RSA groups (see Brakerski et al. TCC 2019).

The proofs of security are very simple, partially due to the fact
that the assumptions are essentially stating that the scheme is secure
(both for the delay encryption and the distributed setup
protocol). This is not a strong negative point, since the same can be
said for the RSW time-lock puzzles. Yet, I feel like this aspect could
be emphasized earlier in the paper and the writeup would benefit from
some additional cautionary warnings.

In terms of writeup, one thing that the paper lacks is a clear
list of the contributions of this work, marking what follows easily
from prior work and what is new from this paper. I think this would
greatly help the paper, as the reader does not have to wait until the
end of the paper to appreciate the contributions of this work.